[FEAT] db 연결 테스트를 위해 프로젝트 관련 컨트롤러, 인터페이스, 레포지토리, 서비스 내용 추가
