+++ src/main/java/com/animalfarm/mlf/domain/project/ProjectInterface.java
@@ -1,0 +1,9 @@
+package com.animalfarm.mlf.domain.project;
+
+import java.util.List;
+
+import com.animalfarm.mlf.domain.project.dto.ProjectDTO;
+
+public interface ProjectInterface {
+	public abstract List<ProjectDTO> selectAll();
+}
