--- src/main/java/com/animalfarrm/mlf/domain/project/ProjectController.java
@@ -1,8 +1,0 @@
-package com.animalfarrm.mlf.domain.project;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ProjectController {
-
-}
