--- src/main/java/com/animalfarrm/mlf/domain/project/ProjectInterface.java
@@ -1,5 +1,0 @@
-package com.animalfarrm.mlf.domain.project;
-
-public interface ProjectInterface {
-
-}
