--- src/main/java/com/animalfarrm/mlf/domain/project/ProjectService.java
@@ -1,8 +1,0 @@
-package com.animalfarrm.mlf.domain.project;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class ProjectService {
-
-}
